Move ScheduledJobQueue into EventLoop and renamed to ScheduledJobs
